refactor: add service layer
