+++ TwitchApi/TwitchAPi.Client/Common/TwitchScopeAttribute.cs
@@ -1,0 +1,9 @@
+﻿namespace TwitchAPi.Client.Common;
+
+[AttributeUsage(AttributeTargets.Field)]
+public class TwitchScopeAttribute : Attribute
+{
+    public string Value { get; }
+    public TwitchScopeAttribute(string value) => Value = value;
+}
+
