+++ TwitchApi/TwitchAPi.Client/TwitchConstants.cs
@@ -1,12 +1,12 @@
-﻿namespace TwitchAPi.Client
-{
-    public static class TwitchConstants
-    {
-        public const string OuathToken = "https://id.twitch.tv/oauth2/token";
-        public const string OuathAuthorize = "https://id.twitch.tv/oauth2/authorize";
-        public const string OauthValidate = "https://id.twitch.tv/oauth2/validate";
+﻿namespace TwitchAPi.Client;
 
-        public const string GetUsers = "https://api.twitch.tv/helix/users";
-        public const string Channels = "https://api.twitch.tv/helix/channels";
-    }
+public static class TwitchConstants
+{
+    public const string OuathToken = "https://id.twitch.tv/oauth2/token";
+    public const string OuathAuthorize = "https://id.twitch.tv/oauth2/authorize";
+    public const string OauthValidate = "https://id.twitch.tv/oauth2/validate";
+
+    public const string GetUsers = "https://api.twitch.tv/helix/users";
+    public const string Channels = "https://api.twitch.tv/helix/channels";
 }
+
