--- TwitchApi/TwitchAPi.Client/TwitchScopeAttribute.cs
@@ -1,9 +1,0 @@
-﻿namespace TwitchAPi.Client
-{
-    [AttributeUsage(AttributeTargets.Field)]
-    public class TwitchScopeAttribute : Attribute
-    {
-        public string Value { get; }
-        public TwitchScopeAttribute(string value) => Value = value;
-    }
-}
